refactor: Rename `Toaster` to `ToastProvider`
