--- src/components/Toaster/index.tsx
@@ -1,1 +1,0 @@
-export { Component as Toaster } from './component';
