+++ src/modules/toast/ToastProvider/index.tsx
@@ -1,0 +1,1 @@
+export { Component as ToastProvider } from './component';
