+++ src/modules/toast/index.tsx
@@ -1,0 +1,1 @@
+export { ToastProvider } from './ToastProvider';
